pyproject.toml
[project]
name = "dataalchemy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.0",
    "openai>=1.93.0",
    "pandas>=2.3.1",
    "pdf2image>=1.17.0",
    "pyarrow>=21.0.0",
    "pydantic-ai>=1.0.17",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
